// src/app/admin/student/model/student-lesson-list.ts
export class StudentLessonList{
    id:number;
    studentName:string;

    name:string;
    code: string;
    stuffId: number;
    departmentId: number;
    semesterId: number;
    credit: number;
    akts: number;
    isSpring: boolean;

    stuffName: string;
    departmentName: string;
    semesterName: string;
}